% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_results.R
\name{get_all_results}
\alias{get_all_results}
\title{Get All Results for a Parkrunner}
\usage{
get_all_results(
  parkrunner,
  folder,
  skip_errors = T,
  log_file = "error_log.txt",
  ...
)
}
\arguments{
\item{parkrunner}{An object of class \code{parkrun_results} (from \code{get_all_runs}).}

\item{folder}{Directory to save CSV files.}

\item{skip_errors}{Logical. If TRUE, skips events listed in \code{log_file}.}

\item{log_file}{Path to the error log file.}

\item{...}{Additional arguments.}
}
\value{
Returns 0 on completion.
}
\description{
Iterates through a parkrunner's history and downloads result files.
}
