% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_runs.R
\name{get_all_runs}
\alias{get_all_runs}
\title{Get All Runs for a Parkrunner}
\usage{
get_all_runs(
  id = NULL,
  url = NULL,
  headers = c(`User-Agent` =
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Safari/537.36",
    Accept =
    "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
    `Accept-Language` = "en-US,en;q=0.9", Connection = "keep-alive")
)
}
\arguments{
\item{id}{The parkrun ID (numeric or string).}

\item{url}{The full URL to the parkrunner's "All Runs" page. If provided, this will not override the \code{id} parameter.}

\item{headers}{A named character vector of HTTP headers.}
}
\value{
A list of class \code{parkrun_results} containing the runner's name, ID, and a data frame of results.
}
\description{
Scrapes the "All Runs" page for a given Parkrunner ID.
}
