% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_history.R
\name{get_event_history}
\alias{get_event_history}
\title{Get Parkrun Results and Volunteer IDs}
\usage{
get_event_history(
  url = NULL,
  event = NULL,
  domain = "parkrun.org.uk",
  headers = c(`User-Agent` =
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Safari/537.36",
    Accept =
    "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
    `Accept-Language` = "en-US,en;q=0.9", Connection = "keep-alive")
)
}
\arguments{
\item{url}{A character string specifying the URL of the parkrun results page. Non-uk prs need a url or update the domain.}

\item{event}{The parkrun event short name (e.g., "bushy"). Required if \code{url} is not provided.}

\item{domain}{The parkrun domain (default is "parkrun.org.uk").}

\item{headers}{A named character vector of HTTP headers to use for the request.}
}
\description{
Scrapes the results table and volunteer IDs from a specific Parkrun event URL.
}
