% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Get Parkrun Results}
\usage{
get_results(
  url,
  headers = c(`User-Agent` =
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Safari/537.36",
    Accept =
    "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
    `Accept-Language` = "en-US,en;q=0.9", Connection = "keep-alive")
)
}
\arguments{
\item{url}{A character string specifying the URL of the parkrun results page.}

\item{headers}{A named character vector of HTTP headers to use for the request.}
}
\value{
A data frame (tibble) containing the position, parkrunner name, and time.
}
\description{
Scrapes the results table from a specific Parkrun event URL.
}
