% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{Get Parkrun Results and Volunteer IDs}
\usage{
get_result(
  url = NULL,
  event = NULL,
  event_no = NULL,
  domain = "parkrun.org.uk",
  headers = c(`User-Agent` =
    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Safari/537.36",
    Accept =
    "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
    `Accept-Language` = "en-US,en;q=0.9", Connection = "keep-alive"),
  as_hms = FALSE,
  as_Date = FALSE,
  timeout = 15,
  extra_data = FALSE
)
}
\arguments{
\item{url}{A character string specifying the URL of the parkrun results page. Non-uk prs need a url or update the domain.}

\item{event}{The parkrun event short name (e.g., "bushy"). Required if \code{url} is not provided.}

\item{event_no}{The parkrun event number (e.g., 1).}

\item{domain}{The parkrun domain (default is "parkrun.org.uk").}

\item{headers}{A named character vector of HTTP headers to use for the request.}

\item{as_hms}{Return times as hms}

\item{as_Date}{Return dates as Date}

\item{timeout}{Control httr timeout}

\item{extra_data}{Extract extra data from results table (currently number of finishes)}
}
\value{
A list containing two elements:
\describe{
\item{results}{A data frame (tibble) with parkrun results (position, parkrunner, time, ag, id).}
\item{volunteers}{A dataframe with volunteer IDs and names.}
\item{date}{The date of the event in ISO format (YYYY-MM-DD).}

}
}
\description{
Scrapes the results table and volunteer IDs from a specific Parkrun event URL.
}
